chore: clean up sudo remove migration file
