+++ runtime/litentry/src/lib.rs
@@ -76,7 +76,6 @@ include!(concat!(env!("OUT_DIR"), "/wasm_binary.rs"));
 
 pub mod asset_config;
 pub mod constants;
-pub mod migration;
 pub mod weights;
 pub mod xcm_config;
 
@@ -127,7 +126,6 @@ pub type Executive = frame_executive::Executive<
 	// It was reverse order before.
 	// See the comment before collation related pallets too.
 	AllPalletsWithSystem,
-	migration::RemoveSudoAndStorage<Runtime>,
 >;
 
 impl_opaque_keys! {
